+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+### 1.3.0
+Number of vertices added
+Total number of meshes added
+Display if a mesh is pickable
+When a mesh name is hovered, the mesh is highlighted in the scene
+Added a wireframe button
+
 ### 1.2.0
 Updated for dark themes.
 Added a button to hide/display a specific mesh